market_data_backend_platform/requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
astroid==4.0.3
black==25.12.0
certifi==2026.1.4
cfgv==3.5.0
click==8.3.1
colorama==0.4.6
coverage==7.13.1
dill==0.4.0
distlib==0.4.0
fastapi==0.128.0
filelock==3.20.2
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==7.0.0
librt==0.7.7
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.4
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
bcrypt==5.0.0
PyJWT==2.10.1
python-multipart==0.0.22
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
SQLAlchemy==2.0.46
starlette==0.50.0
structlog==25.5.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
